+++ Source/RethinkDb.Driver/Model/Result.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Linq;
 
 namespace RethinkDb.Driver.Model
@@ -59,7 +60,12 @@ namespace RethinkDb.Driver.Model
         /// <summary>
         /// If returnChanges is set to true, this will be an array of objects, one for each objected affected by the insert operation. Each object will have two keys: {new_val: -new value-, old_val: null}.
         /// </summary>
-        public JObject Changes { get; set; }
+        public JArray Changes { get; set; }
+
+        public Change<T>[] ChangesAs<T>()
+        {
+            return this.Changes?.ToObject<Change<T>[]>(Net.Converter.Serializer);
+        }
 
 
         /// <summary>
@@ -76,5 +82,34 @@ namespace RethinkDb.Driver.Model
         public uint TablesCreated { get; set; }
         [JsonProperty("tables_dropped")]
         public uint TablesDropped { get; set; }
+    }
+
+    public static class ExtensionsForResult
+    {
+        public static Result EnsureSuccess(this Result result)
+        {
+            if( result.Errors != 0 )
+            {
+                throw new ReqlError(result.FirstError);
+            }
+            return result;
+        }
+    }
+
+    public class Change<T>
+    {
+        [JsonProperty("old_val")]
+        public T OldValue { get; set; }
+        [JsonProperty("new_val")]
+        public T NewValue { get; set; }
+
+        [JsonConverter(typeof(StringEnumConverter))]
+        public ChangeState? State { get; set; }
+    }
+
+    public enum ChangeState
+    {
+        Initializing = 1,
+        Ready
     }
 }